Optimize a bit, indirect drawing still broken

+++ src/render/world/model/static/component/positioned-model-component.ts
@@ -14,6 +14,6 @@ export abstract class PositionedModelComponent implements ModelComponent {
     }
 
     abstract getVertexPositions(parentPosition: Vector3D): Float32Array;
-    abstract getTextureMappings(): Uint32Array;
+    abstract getTextureMappings(): Float32Array;
     abstract getTextureIds(): Uint32Array;
 }
