+++ src/render/world/model/static/model-component.ts
@@ -2,6 +2,6 @@ import { Vector3D } from "../../../../utils/vector3d/vector3d.js";
 
 export interface ModelComponent {
     getVertexPositions(parentPosition: Vector3D): Float32Array;
-    getTextureMappings(): Uint32Array;
+    getTextureMappings(): Float32Array;
     getTextureIds(): Uint32Array;
 }
